+++ eq-author-api/schema/resolvers/utils/folderGetters.js
@@ -1,0 +1,18 @@
+const { find, flatMap, some } = require("lodash");
+const { getSections, getSectionById } = require("./sectionGetters");
+
+const getFolders = (ctx) => flatMap(getSections(ctx), ({ folders }) => folders);
+
+const getFoldersBySectionId = (ctx, id) => getSectionById(ctx, id).folders;
+
+const getFolderById = (ctx, id) => find(getFolders(ctx), { id });
+
+const getFolderByPageId = (ctx, id) =>
+  find(getFolders(ctx), ({ pages }) => pages && some(pages, { id }));
+
+module.exports = {
+  getFolders,
+  getFoldersBySectionId,
+  getFolderById,
+  getFolderByPageId,
+};
