+++ eq-author-api/schema/resolvers/utils/getters.js
@@ -1,79 +1,5 @@
-const { compact, get, filter, find, flatMap, some } = require("lodash");
-
-const getSections = (ctx) => ctx.questionnaire.sections;
-
-const getSectionById = (ctx, id) => find(getSections(ctx), { id });
-
-const getSectionByFolderId = (ctx, folderId) =>
-  find(getSections(ctx), (section) => {
-    if (section.folders && some(section.folders, { id: folderId })) {
-      return section;
-    }
-  });
-
-const getSectionByPageId = (ctx, pageId) =>
-  find(getSections(ctx), (section) =>
-    some(section.folders, (folder) => {
-      if (folder.pages && some(folder.pages, { id: pageId })) {
-        return section;
-      }
-    })
-  );
-
-const getFolders = (ctx) => flatMap(getSections(ctx), ({ folders }) => folders);
-
-const getFoldersBySectionId = (ctx, id) => getSectionById(ctx, id).folders;
-
-const getFolderById = (ctx, id) => find(getFolders(ctx), { id });
-
-const getFolderByPageId = (ctx, id) =>
-  find(getFolders(ctx), ({ pages }) => pages && some(pages, { id }));
-
-const getPages = (ctx) => flatMap(getFolders(ctx), ({ pages }) => pages);
-
-const getPagesBySectionId = (ctx, id) =>
-  flatMap(getSectionById(ctx, id).folders, ({ pages }) => pages);
-
-const getPagesByFolderId = (ctx, id) => getFolderById(ctx, id).pages;
-
-const getPagesFromSection = (section) =>
-  flatMap(section.folders, ({ pages }) => pages);
-
-const getPageById = (ctx, id) => find(getPages(ctx), { id });
-
-const getPageByAnswerId = (ctx, answerId) =>
-  find(
-    getPages(ctx),
-    (page) => page.answers && some(page.answers, { id: answerId })
-  );
-
-const getPageByConfirmationId = (ctx, confirmationId) =>
-  find(getPages(ctx), (page) => {
-    if (get(page, "confirmation.id") === confirmationId) {
-      return page;
-    }
-  });
-
-const getPageByValidationId = (ctx, validationId) =>
-  find(
-    getPages(ctx),
-    (page) => page.totalValidation && page.totalValidation.id === validationId
-  );
-
-const getConfirmations = (ctx) =>
-  compact(flatMap(getPages(ctx), (page) => page.confirmation));
-
-const getConfirmationById = (ctx, id) => find(getConfirmations(ctx), { id });
-
-const getSkippableById = (ctx, id) =>
-  getFolderById(ctx, id) ||
-  getConfirmationById(ctx, id) ||
-  getPageById(ctx, id);
-const getSkippables = (ctx) => [
-  ...getFolders(ctx),
-  ...getConfirmations(ctx),
-  ...getPages(ctx),
-];
+const { compact, filter, find, flatMap, some } = require("lodash");
+const { getPages, getSkipConditions } = require("./pageGetters");
 
 const getAnswers = (ctx) =>
   compact(flatMap(getPages(ctx), (page) => page.answers));
@@ -93,9 +19,6 @@ const getRoutingById = (ctx, id) => find(getRouting(ctx), { id });
 const getRules = (ctx) => flatMap(filter(getRouting(ctx), "rules"), "rules");
 
 const getRoutingRuleById = (ctx, id) => find(getRules(ctx), { id });
-
-const getSkipConditions = (ctx) =>
-  flatMap(filter(getSkippables(ctx), "skipConditions"), "skipConditions");
 
 const getSkipConditionById = (ctx, id) => {
   const skipConditions = getSkipConditions(ctx);
@@ -169,23 +92,23 @@ const getMovePosition = (section, pageId, position) => {
 };
 
 module.exports = {
-  getSections,
-  getSectionById,
-  getSectionByFolderId,
-  getSectionByPageId,
+  // getSections,
+  // getSectionById,
+  // getSectionByFolderId,
+  // getSectionByPageId,
 
-  getFolders,
-  getFoldersBySectionId,
-  getFolderById,
-  getFolderByPageId,
-  getPages,
-  getPagesBySectionId,
-  getPagesByFolderId,
-  getPagesFromSection,
-  getPageById,
-  getPageByConfirmationId,
-  getPageByValidationId,
-  getPageByAnswerId,
+  // getFolders,
+  // getFoldersBySectionId,
+  // getFolderById,
+  // getFolderByPageId,
+  // getPages,
+  // getPagesBySectionId,
+  // getPagesByFolderId,
+  // getPagesFromSection,
+  // getPageById,
+  // getPageByConfirmationId,
+  // getPageByValidationId,
+  // getPageByAnswerId,
 
   getAnswers,
   getAnswerById,
@@ -204,14 +127,14 @@ module.exports = {
   getAllExpressionGroups,
   getExpressionGroupByExpressionId,
 
-  getConfirmations,
-  getConfirmationById,
+  // getConfirmations,
+  // getConfirmationById,
 
-  getSkippableById,
-  getSkippables,
+  // getSkippableById,
+  // getSkippables,
 
   getSkipConditionById,
-  getSkipConditions,
+  // getSkipConditions,
 
   getPosition,
   getMovePosition,
