+++ eq-author-api/schema/resolvers/utils/pageGetters.js
@@ -1,0 +1,69 @@
+const { compact, filter, get, find, flatMap, some } = require("lodash");
+
+const { getFolders, getFolderById } = require("./folderGetters");
+const { getSectionById } = require("./sectionGetters");
+
+const getPages = (ctx) => flatMap(getFolders(ctx), ({ pages }) => pages);
+
+const getPagesBySectionId = (ctx, id) =>
+  flatMap(getSectionById(ctx, id).folders, ({ pages }) => pages);
+
+const getPagesByFolderId = (ctx, id) => getFolderById(ctx, id).pages;
+
+const getPagesFromSection = (section) =>
+  flatMap(section.folders, ({ pages }) => pages);
+
+const getPageById = (ctx, id) => find(getPages(ctx), { id });
+
+const getPageByAnswerId = (ctx, answerId) =>
+  find(
+    getPages(ctx),
+    (page) => page.answers && some(page.answers, { id: answerId })
+  );
+
+const getPageByConfirmationId = (ctx, confirmationId) =>
+  find(getPages(ctx), (page) => {
+    if (get(page, "confirmation.id") === confirmationId) {
+      return page;
+    }
+  });
+
+const getPageByValidationId = (ctx, validationId) =>
+  find(
+    getPages(ctx),
+    (page) => page.totalValidation && page.totalValidation.id === validationId
+  );
+
+const getConfirmations = (ctx) =>
+  compact(flatMap(getPages(ctx), (page) => page.confirmation));
+
+const getConfirmationById = (ctx, id) => find(getConfirmations(ctx), { id });
+
+const getSkippableById = (ctx, id) =>
+  getFolderById(ctx, id) ||
+  getConfirmationById(ctx, id) ||
+  getPageById(ctx, id);
+const getSkippables = (ctx) => [
+  ...getFolders(ctx),
+  ...getConfirmations(ctx),
+  ...getPages(ctx),
+];
+
+const getSkipConditions = (ctx) =>
+  flatMap(filter(getSkippables(ctx), "skipConditions"), "skipConditions");
+
+module.exports = {
+  getPages,
+  getPagesBySectionId,
+  getPagesByFolderId,
+  getPagesFromSection,
+  getPageById,
+  getPageByConfirmationId,
+  getPageByValidationId,
+  getPageByAnswerId,
+  getConfirmations,
+  getConfirmationById,
+  getSkippableById,
+  getSkippables,
+  getSkipConditions,
+};
